test backpressure for response side works

+++ r2-int-test/src/test/java/test/r2/integ/TimedBytesWriter.java
@@ -10,7 +10,7 @@ class TimedBytesWriter extends BytesWriter
   private long _startTime;
   private long _stopTime;
 
-  TimedBytesWriter(int total, byte fill)
+  TimedBytesWriter(long total, byte fill)
   {
     super(total, fill);
   }
